购物车商品数量加减

+++ xm_project/xm_server/routes/cart.js
@@ -22,12 +22,13 @@ router.get('/cart_info',function(req,res){
 })
 
 //删除购物车内容
-//http://localhost:4000/cart/cart_del/?eid=1
+//http://localhost:4000/cart/cart_del?eid=1
 router.get('/cart_del',function(req,res){
   //获取商品的eid
   var $eid=req.query.eid
-  var sql="delete from cart where eid=?";
-  pool.query(sql,[$eid],function(err,result){
+  var uid=req.session.uid
+  var sql="delete from cart where eid=? and uid=?";
+  pool.query(sql,[$eid,uid],function(err,result){
     if(err) throw err
     console.log(result)
     if(result.affectedRows>0){
@@ -37,7 +38,36 @@ router.get('/cart_del',function(req,res){
     }
   })
 })
-
+//购物车产品数量增1
+router.get('/cart_plus',function(req,res){
+  var eid=req.query.eid
+  var uid=req.session.uid
+  var i=req.query.i
+  var sql=`update cart set ecount=ecount+1 where eid=${eid} and uid=${uid}`;
+  pool.query(sql,function(err,result){
+    if(err) throw err
+    if(result.affectedRows>0){
+      res.send({code:1,msg:"增加成功"})
+    }else{
+      res.send({code:-3,msg:"增加失败"})
+    }
+  })
+})
+//购物车产品数量减1
+router.get('/cart_minus',function(req,res){
+  var eid=req.query.eid
+  var uid=req.session.uid
+  var i=req.query.i
+  var sql=`update cart set ecount=ecount-1 where eid=${eid} and uid=${uid}`;
+  pool.query(sql,function(err,result){
+    if(err) throw err
+    if(result.affectedRows>0){
+      res.send({code:1,msg:"减少成功"})
+    }else{
+      res.send({code:-3,msg:"减少失败"})
+    }
+  })
+})
 
 
 //导出路由
